Complete migration to PHPUnit 5/6 with duplicated legacy classes
