+++ src/Test/PHPUnit/AbstractCacheConstraint.php
@@ -11,71 +11,18 @@
 
 namespace FOS\HttpCache\Test\PHPUnit;
 
-use Psr\Http\Message\ResponseInterface;
+use PHPUnit\Framework\Constraint;
 
-/**
- * Abstract cache constraint.
- */
-abstract class AbstractCacheConstraint extends \PHPUnit_Framework_Constraint
-{
-    protected $header = 'X-Cache';
-
+if (class_exists('PHPUnit_Runner_Version') && version_compare(\PHPUnit_Runner_Version::id(), '6.0.0', '<')) {
+    class_alias('FOS\HttpCache\Test\PHPUnit\AbstractCacheConstraint', 'FOS\HttpCache\Test\Legacy\PHPUnit\AbstractCacheConstraint');
+// Using an early return instead of a else does not work when using the PHPUnit phar due to some weird PHP behavior (the class
+// gets defined without executing the code before it and so the definition is not properly conditional)
+} else {
     /**
-     * Constructor.
-     *
-     * @param string $header Cache debug header; defaults to X-Cache-Debug
+     * Abstract cache constraint.
      */
-    public function __construct($header = null)
+    abstract class AbstractCacheConstraint extends Constraint
     {
-        if ($header) {
-            $this->header = $header;
-        }
-
-        parent::__construct();
-    }
-
-    /**
-     * Get cache header value.
-     *
-     * @return string
-     */
-    abstract public function getValue();
-
-    /**
-     * {@inheritdoc}
-     *
-     * @param Response $other The guzzle response object
-     */
-    protected function matches($other)
-    {
-        if (!$other instanceof ResponseInterface) {
-            throw new \RuntimeException(sprintf('Expected a GuzzleHttp\Psr7\Response but got %s', get_class($other)));
-        }
-        if (!$other->hasHeader($this->header)) {
-            $message = sprintf(
-                'Response has no "%s" header. Configure your caching proxy '
-                .'to set the header with cache hit/miss status.',
-                $this->header
-            );
-            if (200 !== $other->getStatusCode()) {
-                $message .= sprintf("\nStatus code of response is %s.", $other->getStatusCode());
-            }
-
-            throw new \RuntimeException($message);
-        }
-
-        return strpos((string) $other->getHeaderLine($this->header), $this->getValue()) !== false;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function failureDescription($other)
-    {
-        return sprintf(
-            'response (with status code %s) %s',
-            $other->getStatusCode(),
-            $this->toString()
-        );
+        use AbstractCacheConstraintTrait;
     }
 }
