+++ src/Test/WebServerListener.php
@@ -11,179 +11,141 @@
 
 namespace FOS\HttpCache\Test;
 
-/**
- * A PHPUnit test listener that starts and stops the PHP built-in web server.
- *
- * This listener is configured with a couple of constants from the phpunit.xml
- * file. To define constants in the phpunit file, use this syntax:
- * <php>
- *     <const name="WEB_SERVER_HOSTNAME" value="localhost" />
- * </php>
- *
- * WEB_SERVER_HOSTNAME host name of the web server (required)
- * WEB_SERVER_PORT     port to listen on (required)
- * WEB_SERVER_DOCROOT  path to the document root for the server (required)
- */
-class WebServerListener implements \PHPUnit_Framework_TestListener
-{
-    /**
-     * PHP web server PID.
-     *
-     * @var int
-     */
-    protected $pid;
+use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\Test;
+use PHPUnit\Framework\TestListener;
+use PHPUnit\Framework\TestSuite;
 
+if (class_exists('PHPUnit_Runner_Version') && version_compare(\PHPUnit_Runner_Version::id(), '6.0.0', '<')) {
+    class_alias('FOS\HttpCache\Test\WebServerListener', 'FOS\HttpCache\Test\Legacy\WebServerListener');
+// Using an early return instead of a else does not work when using the PHPUnit phar due to some weird PHP behavior (the class
+// gets defined without executing the code before it and so the definition is not properly conditional)
+} else {
     /**
-     * Make sure the PHP built-in web server is running for tests with group
-     * 'webserver'.
+     * A PHPUnit test listener that starts and stops the PHP built-in web server.
+     *
+     * This listener is configured with a couple of constants from the phpunit.xml
+     * file. To define constants in the phpunit file, use this syntax:
+     * <php>
+     *     <const name="WEB_SERVER_HOSTNAME" value="localhost" />
+     * </php>
+     *
+     * WEB_SERVER_HOSTNAME host name of the web server (required)
+     * WEB_SERVER_PORT     port to listen on (required)
+     * WEB_SERVER_DOCROOT  path to the document root for the server (required)
      */
-    public function startTestSuite(\PHPUnit_Framework_TestSuite $suite)
+    class WebServerListener implements TestListener
     {
-        // Only run on PHP >= 5.4 as PHP below that and HHVM don't have a
-        // built-in web server
-        if (defined('HHVM_VERSION')) {
-            return;
+        /** @var WebServerListenerTrait */
+        private $trait;
+
+        public function __construct()
+        {
+            $this->trait = new WebServerListenerTrait();
         }
 
-        if (!in_array('webserver', $suite->getGroups()) || null !== $this->pid) {
-            return;
+        /**
+         * Make sure the PHP built-in web server is running for tests with group
+         * 'webserver'.
+         */
+        public function startTestSuite(TestSuite $suite)
+        {
+            $this->trait->startTestSuite($suite);
         }
 
-        $this->pid = $pid = $this->startPhpWebServer();
-
-        register_shutdown_function(function () use ($pid) {
-            exec('kill '.$pid);
-        });
-    }
-
-    /**
-     *  We don't need these.
-     */
-    public function endTestSuite(\PHPUnit_Framework_TestSuite $suite)
-    {
-    }
-
-    public function addError(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addFailure(\PHPUnit_Framework_Test $test, \PHPUnit_Framework_AssertionFailedError $e, $time)
-    {
-    }
-
-    public function addIncompleteTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addSkippedTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function startTest(\PHPUnit_Framework_Test $test)
-    {
-    }
-
-    public function endTest(\PHPUnit_Framework_Test $test, $time)
-    {
-    }
-
-    public function addRiskyTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    /**
-     * Get web server hostname.
-     *
-     * @throws \Exception
-     *
-     * @return string
-     */
-    protected function getHostName()
-    {
-        if (!defined('WEB_SERVER_HOSTNAME')) {
-            throw new \Exception('Set WEB_SERVER_HOSTNAME in your phpunit.xml');
+        /**
+         *  We don't need these.
+         */
+        public function endTestSuite(TestSuite $suite)
+        {
         }
 
-        return WEB_SERVER_HOSTNAME;
-    }
-
-    /**
-     * Get web server port.
-     *
-     * @throws \Exception
-     *
-     * @return int
-     */
-    protected function getPort()
-    {
-        if (!defined('WEB_SERVER_PORT')) {
-            throw new \Exception('Set WEB_SERVER_PORT in your phpunit.xml');
+        public function addError(Test $test, \Exception $e, $time)
+        {
         }
 
-        return WEB_SERVER_PORT;
-    }
-
-    /**
-     * Get web server port.
-     *
-     * @throws \Exception
-     *
-     * @return int
-     */
-    protected function getDocRoot()
-    {
-        if (!defined('WEB_SERVER_DOCROOT')) {
-            throw new \Exception('Set WEB_SERVER_DOCROOT in your phpunit.xml');
+        public function addFailure(Test $test, AssertionFailedError $e, $time)
+        {
         }
 
-        return WEB_SERVER_DOCROOT;
-    }
-
-    /**
-     * Start PHP built-in web server.
-     *
-     * @return int PID
-     */
-    protected function startPhpWebServer()
-    {
-        $command = sprintf(
-            'php -S %s:%d -t %s >/dev/null 2>&1 & echo $!',
-            '127.0.0.1', // on travis, localhost is not 127.0.0.1 but IPv6 ::1
-            $this->getPort(),
-            $this->getDocRoot()
-        );
-        exec($command, $output);
-
-        $this->waitFor($this->getHostName(), $this->getPort(), 2000);
-
-        return $output[0];
-    }
-
-    /**
-     * Wait for caching proxy to be started up and reachable.
-     *
-     * @param string $ip
-     * @param int    $port
-     * @param int    $timeout Timeout in milliseconds
-     *
-     * @throws \RuntimeException If proxy is not reachable within timeout
-     */
-    protected function waitFor($ip, $port, $timeout)
-    {
-        for ($i = 0; $i < $timeout; ++$i) {
-            if (@fsockopen($ip, $port)) {
-                return;
-            }
-
-            usleep(1000);
+        public function addIncompleteTest(Test $test, \Exception $e, $time)
+        {
         }
 
-        throw new \RuntimeException(
-            sprintf(
-                'Webserver cannot be reached at %s:%s',
-                $ip,
-                $port
-            )
-        );
+        public function addSkippedTest(Test $test, \Exception $e, $time)
+        {
+        }
+
+        public function startTest(Test $test)
+        {
+        }
+
+        public function endTest(Test $test, $time)
+        {
+        }
+
+        public function addRiskyTest(Test $test, \Exception $e, $time)
+        {
+        }
+
+        /**
+         * Get web server hostname.
+         *
+         * @throws \Exception
+         *
+         * @return string
+         */
+        protected function getHostName()
+        {
+            return $this->trait->getHostName();
+        }
+
+        /**
+         * Get web server port.
+         *
+         * @throws \Exception
+         *
+         * @return int
+         */
+        protected function getPort()
+        {
+            return $this->trait->getPort();
+        }
+
+        /**
+         * Get web server port.
+         *
+         * @throws \Exception
+         *
+         * @return int
+         */
+        protected function getDocRoot()
+        {
+            return $this->trait->getDocRoot();
+        }
+
+        /**
+         * Start PHP built-in web server.
+         *
+         * @return int PID
+         */
+        protected function startPhpWebServer()
+        {
+            return $this->trait->startPhpWebServer();
+        }
+
+        /**
+         * Wait for caching proxy to be started up and reachable.
+         *
+         * @param string $ip
+         * @param int $port
+         * @param int $timeout Timeout in milliseconds
+         *
+         * @throws \RuntimeException If proxy is not reachable within timeout
+         */
+        protected function waitFor($ip, $port, $timeout)
+        {
+            $this->trait->waitFor($ip, $port, $timeout);
+        }
     }
 }
